+++ Web06/src/spms/listeners/ContextLoaderListener.java
@@ -8,7 +8,13 @@ import javax.servlet.ServletContextListener;
 import javax.servlet.annotation.WebListener;
 import javax.sql.DataSource;
 
-import spms.dao.MemberDao;
+import smps.controls.LoginController;
+import smps.controls.LogoutController;
+import smps.controls.MemberAddController;
+import smps.controls.MemberDeleteController;
+import smps.controls.MemberListController;
+import smps.controls.MemberUpdateController;
+import spms.dao.MySqlMemberDao;
 
 @WebListener
 public class ContextLoaderListener implements ServletContextListener {
@@ -21,11 +27,15 @@ public class ContextLoaderListener implements ServletContextListener {
       DataSource ds = (DataSource)initialContext.lookup(
           "java:comp/env/jdbc/studydb");
       
-      MemberDao memberDao = new MemberDao();
+      MySqlMemberDao memberDao = new MySqlMemberDao();
       memberDao.setDataSource(ds);
       
-      sc.setAttribute("memberDao", memberDao);
-
+      sc.setAttribute("/auth/login.do", new LoginController().setMemberDao(memberDao));
+      sc.setAttribute("/auth/logout.do", new LogoutController());
+      sc.setAttribute("/member/list.do", new MemberListController().setMemberDao(memberDao));
+      sc.setAttribute("/member/add.do", new MemberAddController().setMemberDao(memberDao));
+      sc.setAttribute("/member/update.do", new MemberUpdateController().setMemberDao(memberDao));
+      sc.setAttribute("/member/delete.do", new MemberDeleteController().setMemberDao(memberDao));
     } catch(Throwable e) {
       e.printStackTrace();
     }
